-Wsign-compare warnings in FTLLowerDFGToB3.cpp and DFGSpeculativeJIT.cpp https://bugs.webkit.org/show_bug.cgi?id=211783
Patch by Michael Catanzaro <mcatanzaro@gnome.org> on 2020-05-12
Reviewed by Darin Adler.

This fixes -Wsign-compare warnings introduced in r260331.

* dfg/DFGSpeculativeJIT.cpp:
(JSC::DFG::SpeculativeJIT::compileValueBitNot):
* ftl/FTLLowerDFGToB3.cpp:
(JSC::FTL::DFG::LowerDFGToB3::compileValueBitNot):

Canonical link: https://commits.webkit.org/224679@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@261564 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/JavaScriptCore/dfg/DFGSpeculativeJIT.cpp
@@ -3597,7 +3597,7 @@ void SpeculativeJIT::compileValueBitNot(Node* node)
 
         // The following trick relies on details of the representation of BigInt32, and will have to be updated if we move bits around.
         static_assert(JSValue::BigInt32Tag == 0x12);
-        static_assert(JSValue::BigInt32Mask == 0xfffe000000000012);
+        static_assert(JSValue::BigInt32Mask == static_cast<int64_t>(0xfffe000000000012));
         constexpr uint64_t maskForBigInt32Bits = 0x0000ffffffff0000;
         static_assert(!(JSValue::BigInt32Mask & maskForBigInt32Bits));
         m_jit.move(TrustedImm64(maskForBigInt32Bits), resultGPR);
